Reorganization + added control surfaces to Foil + prototype
